+++ src/art/lib/ArtJobExecutor.cc
@@ -57,14 +57,12 @@ namespace {
 
         DELETE_COPY(AJEReporter)
         DELETE_MOVE(AJEReporter)
-        ~AJEReporter(){
-            stop();
-        }
+        ~AJEReporter() { stop(); }
 
         void stop()
         {
             should_stop_ = true;
-            if (thread_.joinable()){
+            if (thread_.joinable()) {
                 thread_.join();
             }
         }
@@ -215,7 +213,8 @@ void ArtJobExecutor::operator()()
             num_pos_reads = npr;
             num_neg_reads = nnr;
         } else {
-            const auto [npr, nnr] = calculate_num_reads_old(contig_size, art_params_.read_len_1, cov_pos, cov_neg, num_reads_to_reduce_);
+            const auto [npr, nnr]
+                = calculate_num_reads_old(contig_size, art_params_.read_len_1, cov_pos, cov_neg, num_reads_to_reduce_);
             num_pos_reads = npr;
             num_neg_reads = nnr;
         }