+++ src/cmake.d/rng.cmake
@@ -21,11 +21,23 @@ elseif("${USE_RANDOM_GENERATOR}" STREQUAL "SYSTEM_PCG")
     set(USE_SYSTEM_PCG_RANDOM ON)
     # TODO: Find <pcg_random.hpp>
 elseif("${USE_RANDOM_GENERATOR}" STREQUAL "ONEMKL")
-    find_package(MKL REQUIRED)
-    set(USE_ONEMKL_RANDOM ON)
-    set(WITH_ONEMKL ON)
-    include_directories(${MKL_INCLUDE} ${SYCL_INCLUDE_DIR})
-    set(ART_MODERN_LINK_LIBS ${ART_MODERN_LINK_LIBS} MKL::MKL)
+    if(DEFINED FIND_RANDOM_MKL_THROUGH_PKGCONF)
+        ceu_cm_enhanced_find_library(OUTPUT_VARIABLE mkl LINKER_FLAG "${FIND_RANDOM_MKL_THROUGH_PKGCONF}" PKGCONF_NAME
+                                     "${FIND_RANDOM_MKL_THROUGH_PKGCONF}")
+
+        if(NOT TARGET CEU_CM_EFL::mkl)
+            message(FATAL_ERROR "mkl (${FIND_RANDOM_MKL_THROUGH_PKGCONF}) not found!")
+        endif()
+        set(USE_ONEMKL_RANDOM ON)
+        set(WITH_ONEMKL ON)
+        set(ART_MODERN_LINK_LIBS ${ART_MODERN_LINK_LIBS} "CEU_CM_EFL::mkl")
+    else()
+        find_package(MKL REQUIRED)
+        set(USE_ONEMKL_RANDOM ON)
+        set(WITH_ONEMKL ON)
+        include_directories(${MKL_INCLUDE} ${SYCL_INCLUDE_DIR})
+        set(ART_MODERN_LINK_LIBS ${ART_MODERN_LINK_LIBS} MKL::MKL)
+    endif()
 else()
     message(
         FATAL_ERROR